corrected minor typos and moved const declarations [JsonRpcRequestExample.scala]
